requirements.txt
mediapipe
opencv-python
pywebview